ModV module refactoring WIP: Layer, messageHandler
